pyproject.toml
[project]
name = "story-crafter-adk"
version = "0.1.0"
description = "Multi-agent story generation system powered by the Google Agent Development Kit (ADK-only, no memory implementation)."
authors = [{ name = "Story Crafter Contributors" }]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.121.2",
    "google-adk>=0.2.0",
    "pydantic>=2.9.0",
    "python-dotenv>=1.0.1",
    "uvicorn>=0.38.0",
]

[project.urls]
Homepage = "https://github.com/StoryCrafter/story-crafter-adk"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["agents*", "models*", "orchestrator*"]

[dependency-groups]
dev = [
    "pytest>=8.3.2",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.7.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
